Refactoring file handling utilities
