+++ azureml/components/src/shared/jsonl_utils.py
@@ -4,54 +4,30 @@ import json
 import pathlib
 import tempfile
 
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from typing import Any, Callable, Tuple
 
+from .jsonl_file_utils import JSONLReader
 from .logging_utils import get_standard_logger_for_file
 
 _logger = get_standard_logger_for_file(__file__)
 
 
-def load_jsonl(file_path: pathlib.Path, source_encoding: str) -> List[Dict[str, Any]]:
-    result = []
-    _logger.info(f"Loading JSON file: {file_path}")
-    with open(file_path, "r", encoding=source_encoding) as jlf:
-        current_line = 0
-        for l in jlf:
-            _logger.info(f"Processing line: {current_line}")
-            nxt = json.loads(l)
-            result.append(nxt)
-            current_line += 1
-    return result
-
-
-def save_jsonl(
-    file_path: pathlib.Path, data: List[Dict[str, Any]], destination_encoding: str
-):
-    _logger.info(f"Saving file {file_path}")
-    with open(file_path, "w", encoding=destination_encoding) as out_file:
-        for i, d in enumerate(data):
-            _logger.info(f"Writing element {i}")
-            d_str = json.dumps(d)
-            out_file.write(d_str)
-            out_file.write("\n")
-
-
 def line_map(
     *,
-    map_func: Callable[[Dict[str, Any]], Optional[Dict[str, Any]]],
+    map_func: Callable[[dict[str, Any]], dict[str, Any] | None],
     source_file: pathlib.Path,
     dest_file: pathlib.Path,
     source_encoding: str,
     dest_encoding: str,
-    error_file: Optional[pathlib.Path] = None,
-    error_encoding: Optional[str] = None,
+    error_file: pathlib.Path | None = None,
+    error_encoding: str | None = None,
     max_errors: int = -1,
 ) -> Tuple[int, int]:
     """Iterate over a JSONL file, applying map_func to each line"""
     assert source_file.exists()
 
     # If error_file is not specified, set up a temporary file
-    def get_error_file(error_file_path: Optional[pathlib.Path]):
+    def get_error_file(error_file_path: pathlib.Path | None):
         if error_file_path:
             return open(error_file_path, "a", encoding=error_encoding)
         else:
@@ -59,15 +35,14 @@ def line_map(
 
     successful_lines = 0
     error_lines = 0
-    with open(source_file, "r", encoding=source_encoding) as in_file:
+    with JSONLReader(source_file, source_encoding) as in_file:
         with open(dest_file, "w", encoding=dest_encoding) as out_file:
             with get_error_file(error_file) as err_file:
                 current_line = 0
                 for nxt in in_file:
                     _logger.info(f"Processing line: {current_line}")
-                    nxt_dict = json.loads(nxt)
                     try:
-                        nxt_output = map_func(nxt_dict)
+                        nxt_output = map_func(nxt)
                         if nxt_output is not None:
                             nxt_output_string = json.dumps(nxt_output)
                             _logger.info(f"Writing output: {nxt_output_string}")
@@ -78,7 +53,7 @@ def line_map(
                         successful_lines += 1
                     except Exception as e:
                         _logger.warn(f"Caught exception: {e}")
-                        err_file.write(nxt)
+                        err_file.write(json.dumps(nxt))
                         error_lines += 1
                     current_line += 1
 
@@ -92,18 +67,17 @@ def line_map(
 
 def line_reduce(
     *,
-    reducer: Callable[[Dict[str, Any]], None],
+    reducer: Callable[[dict[str, Any]], None],
     source_file: pathlib.Path,
     source_encoding: str,
 ):
     assert source_file.exists()
 
-    with open(source_file, "r", encoding=source_encoding) as in_file:
+    with JSONLReader(source_file, source_encoding) as in_file:
         current_line = 0
         for nxt in in_file:
             _logger.info(f"Processing line: {current_line}")
             current_line += 1
-            nxt_dict = json.loads(nxt)
             _logger.info(f"Calling reducer")
-            reducer(nxt_dict)
+            reducer(nxt)
     _logger.info(f"line_reduce complete")
